fix(days-of-supply): PR changes
