+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/delivery/domain/repository/DeliveryRepository.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2024 Volkswagen AG
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023, 2024 Volkswagen AG
+ * Copyright (c) 2023, 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -18,18 +18,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.eclipse.tractusx.puris.backend.supply.domain.model;
+package org.eclipse.tractusx.puris.backend.delivery.domain.repository;
 
-import jakarta.persistence.Entity;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-import lombok.experimental.SuperBuilder;
+import java.util.UUID;
 
-@Entity
-@SuperBuilder
-@NoArgsConstructor
-@ToString(callSuper = true)
-public class OwnSupplierSupply extends Supply {
+import org.eclipse.tractusx.puris.backend.delivery.domain.model.Delivery;
+import org.springframework.data.jpa.repository.JpaRepository;
 
+public interface DeliveryRepository<T extends Delivery> extends JpaRepository<T, UUID> {
+    
 }
-
