+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/delivery/domain/repository/OwnDeliveryRepository.java
@@ -20,11 +20,8 @@
 
 package org.eclipse.tractusx.puris.backend.delivery.domain.repository;
 
-import java.util.UUID;
-
 import org.eclipse.tractusx.puris.backend.delivery.domain.model.OwnDelivery;
-import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OwnDeliveryRepository extends JpaRepository<OwnDelivery, UUID> {
+public interface OwnDeliveryRepository extends DeliveryRepository<OwnDelivery> {
     
 }
