+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/delivery/logic/service/ReportedDeliveryService.java
@@ -20,28 +20,20 @@
 
 package org.eclipse.tractusx.puris.backend.delivery.logic.service;
 
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneId;
-import java.time.ZoneOffset;
-import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.function.Function;
-import java.util.stream.Stream;
 
 import org.eclipse.tractusx.puris.backend.delivery.domain.model.EventTypeEnumeration;
 import org.eclipse.tractusx.puris.backend.delivery.domain.model.ReportedDelivery;
 import org.eclipse.tractusx.puris.backend.delivery.domain.repository.ReportedDeliveryRepository;
 import org.eclipse.tractusx.puris.backend.masterdata.domain.model.Partner;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.service.PartnerService;
-import org.eclipse.tractusx.puris.backend.stock.logic.dto.itemstocksamm.DirectionCharacteristic;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ReportedDeliveryService {
+public class ReportedDeliveryService extends DeliveryService<ReportedDelivery> {
     private final ReportedDeliveryRepository repository;
 
     private final PartnerService partnerService;
@@ -56,80 +48,9 @@ public class ReportedDeliveryService {
         this.validator = this::validate;
     }
 
-    public final List<ReportedDelivery> findAll() {
-        return repository.findAll();
-    }
-
     public final List<ReportedDelivery> findAllByReportedId(UUID reportedId) {
         return repository.findAll().stream().filter(delivery -> delivery.getPartner().getUuid().equals(reportedId))
             .toList();
-    }
-
-    public final ReportedDelivery findById(UUID id) {
-        return repository.findById(id).orElse(null);
-    }
-
-    public final List<ReportedDelivery> findAllByFilters(
-        Optional<String> ownMaterialNumber,
-        Optional<String> bpns,
-        Optional<String> bpnl,
-        Optional<Date> day,
-        Optional<DirectionCharacteristic> direction) {
-        Stream<ReportedDelivery> stream = repository.findAll().stream();
-        if (ownMaterialNumber.isPresent()) {
-            stream = stream.filter(delivery -> delivery.getMaterial().getOwnMaterialNumber().equals(ownMaterialNumber.get()));
-        }
-        if (bpns.isPresent()) {
-            stream = stream.filter(delivery -> delivery.getDestinationBpns().equals(bpns.get()) || delivery.getOriginBpns().equals(bpns.get()));
-        }
-        if (bpnl.isPresent()) {
-            stream = stream.filter(delivery -> delivery.getPartner().getBpnl().equals(bpnl.get()));
-        }
-        if (day.isPresent()) {
-            LocalDate localDayDate = Instant.ofEpochMilli(day.get().getTime())
-                .atOffset(ZoneOffset.UTC)
-                .toLocalDate();
-            stream = stream.filter(delivery -> {
-                long time = direction.get() == DirectionCharacteristic.INBOUND
-                    ? delivery.getDateOfArrival().getTime()
-                    : delivery.getDateOfDeparture().getTime();
-                LocalDate deliveryDayDate = Instant.ofEpochMilli(time)
-                    .atOffset(ZoneOffset.UTC)
-                    .toLocalDate();
-                return deliveryDayDate.getDayOfMonth() == localDayDate.getDayOfMonth();
-            });
-        }
-        if (direction.isPresent()) {
-            if (direction.get() == DirectionCharacteristic.INBOUND) {
-                stream = stream.filter(delivery -> delivery.getDestinationBpns().equals(bpns.get()));
-            } else {
-                stream = stream.filter(delivery -> delivery.getOriginBpns().equals(bpns.get()));
-            }
-        }
-        return stream.toList();
-    }
-
-    public final double getSumOfQuantities(List<ReportedDelivery> deliveries) {
-        double sum = 0;
-        for (ReportedDelivery delivery : deliveries) {
-            sum += delivery.getQuantity();
-        }
-        return sum;
-    }
-
-    public final List<Double> getQuantityForDays(String material, String partnerBpnl, String siteBpns, DirectionCharacteristic direction, int numberOfDays) {
-        List<Double> deliveryQtys = new ArrayList<>();
-        LocalDate localDate = LocalDate.now();
-
-        for (int i = 0; i < numberOfDays; i++) {
-            Date date = Date.from(localDate.atStartOfDay(ZoneId.systemDefault()).toInstant());
-            List<ReportedDelivery> deliveries = findAllByFilters(Optional.of(material), Optional.of(siteBpns), Optional.of(partnerBpnl), Optional.of(date), Optional.of(direction));
-            double deliveryQuantity = getSumOfQuantities(deliveries);
-            deliveryQtys.add(deliveryQuantity);
-
-            localDate = localDate.plusDays(1);
-        }
-        return deliveryQtys;
     }
 
     public final ReportedDelivery create(ReportedDelivery delivery) {
@@ -151,17 +72,6 @@ public class ReportedDeliveryService {
             return null;
         }
         return repository.saveAll(deliveries);
-    }
-
-    public final ReportedDelivery update(ReportedDelivery delivery) {
-        if (delivery.getUuid() == null || repository.findById(delivery.getUuid()).isEmpty()) {
-            return null;
-        }
-        return repository.save(delivery);
-    }
-
-    public final void delete(UUID id) {
-        repository.deleteById(id);
     }
 
     public boolean validate(ReportedDelivery delivery) {
