+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/production/domain/repository/OwnProductionRepository.java
@@ -21,12 +21,9 @@ SPDX-License-Identifier: Apache-2.0
 package org.eclipse.tractusx.puris.backend.production.domain.repository;
 
 import org.eclipse.tractusx.puris.backend.production.domain.model.OwnProduction;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface OwnProductionRepository extends JpaRepository<OwnProduction, UUID> {
+public interface OwnProductionRepository extends ProductionRepository<OwnProduction> {
 
 }
