+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/supply/domain/model/OwnCustomerSupply.java
@@ -1,35 +1,35 @@
 /*
- * Copyright (c) 2024 Volkswagen AG
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
+* Copyright (c) 2024 Volkswagen AG
+* Copyright (c) 2024 Contributors to the Eclipse Foundation
+*
+* See the NOTICE file(s) distributed with this work for additional
+* information regarding copyright ownership.
+*
+* This program and the accompanying materials are made available under the
+* terms of the Apache License, Version 2.0 which is available at
+* https://www.apache.org/licenses/LICENSE-2.0.
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+* WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+* License for the specific language governing permissions and limitations
+* under the License.
+*
+* SPDX-License-Identifier: Apache-2.0
+*/
 
- package org.eclipse.tractusx.puris.backend.supply.domain.model;
+package org.eclipse.tractusx.puris.backend.supply.domain.model;
 
- import jakarta.persistence.Entity;
- import lombok.NoArgsConstructor;
- import lombok.ToString;
- import lombok.experimental.SuperBuilder;
- 
- @Entity
- @SuperBuilder
- @NoArgsConstructor
- @ToString(callSuper = true)
- public class OwnCustomerSupply extends Supply {
- 
- }
- 
+import jakarta.persistence.Entity;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+
+@Entity
+@SuperBuilder
+@NoArgsConstructor
+@ToString(callSuper = true)
+public class OwnCustomerSupply extends Supply {
+
+}
+
