+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/supply/domain/repository/ReportedCustomerSupplyRepository.java
@@ -20,11 +20,12 @@
 
 package org.eclipse.tractusx.puris.backend.supply.domain.repository;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.eclipse.tractusx.puris.backend.supply.domain.model.ReportedCustomerSupply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportedCustomerSupplyRepository extends JpaRepository<ReportedCustomerSupply, UUID> {
-    
+    List<ReportedCustomerSupply> findByPartner_BpnlAndMaterial_OwnMaterialNumber(String partnerBpnl, String ownMaterialNumber);
 }
