+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/supply/domain/repository/ReportedSupplierSupplyRepository.java
@@ -20,11 +20,13 @@
 
 package org.eclipse.tractusx.puris.backend.supply.domain.repository;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.eclipse.tractusx.puris.backend.supply.domain.model.ReportedSupplierSupply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReportedSupplierSupplyRepository extends JpaRepository<ReportedSupplierSupply, UUID> {
-    
+    List<ReportedSupplierSupply> findByPartner_BpnlAndMaterial_OwnMaterialNumber(String partnerBpnl, String ownMaterialNumber);
 }
+
