+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/supply/logic/service/SupplierSupplyService.java
@@ -32,7 +32,7 @@ import java.util.stream.Stream;
 
 import org.eclipse.tractusx.puris.backend.delivery.logic.service.OwnDeliveryService;
 import org.eclipse.tractusx.puris.backend.delivery.logic.service.ReportedDeliveryService;
-import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialServiceImpl;
+import org.eclipse.tractusx.puris.backend.masterdata.logic.service.MaterialService;
 import org.eclipse.tractusx.puris.backend.masterdata.logic.service.PartnerService;
 import org.eclipse.tractusx.puris.backend.production.logic.service.OwnProductionService;
 import org.eclipse.tractusx.puris.backend.stock.logic.dto.itemstocksamm.DirectionCharacteristic;
@@ -40,39 +40,43 @@ import org.eclipse.tractusx.puris.backend.stock.logic.service.MaterialItemStockS
 import org.eclipse.tractusx.puris.backend.supply.domain.model.OwnSupplierSupply;
 import org.eclipse.tractusx.puris.backend.supply.domain.model.ReportedSupplierSupply;
 import org.eclipse.tractusx.puris.backend.supply.domain.repository.ReportedSupplierSupplyRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
 public class SupplierSupplyService {
-    private final ReportedSupplierSupplyRepository repository; 
-    private final PartnerService partnerService;
-    private final OwnDeliveryService ownDeliveryService;
-    private final ReportedDeliveryService reportedDeliveryService;
-    private final OwnProductionService productionService;
-    private final MaterialItemStockService stockService;
-    private final MaterialServiceImpl materialService;
+    @Autowired
+    private ReportedSupplierSupplyRepository repository; 
+    @Autowired
+    private PartnerService partnerService;
+    @Autowired
+    private OwnDeliveryService ownDeliveryService;
+    @Autowired
+    private ReportedDeliveryService reportedDeliveryService;
+    @Autowired
+    private OwnProductionService productionService;
+    @Autowired
+    private MaterialItemStockService stockService;
+    @Autowired
+    private MaterialService materialService;
 
     protected final Function<ReportedSupplierSupply, Boolean> validator;
 
-    public SupplierSupplyService(
-        ReportedSupplierSupplyRepository repository,
-        PartnerService partnerService,
-        OwnDeliveryService ownDeliveryService,
-        ReportedDeliveryService reportedDeliveryService,
-        OwnProductionService productionService,
-        MaterialItemStockService stockService,
-        MaterialServiceImpl materialService) {
-        this.repository = repository;
-        this.partnerService = partnerService;
-        this.ownDeliveryService = ownDeliveryService;
-        this.reportedDeliveryService = reportedDeliveryService;
-        this.productionService = productionService;
-        this.stockService = stockService;
-        this.materialService = materialService;
+    public SupplierSupplyService() {
         this.validator = this::validate;
     }
 
-        public final List<OwnSupplierSupply> calculateSupplierDaysOfSupply(String material, String partnerBpnl, String siteBpns, int numberOfDays) {
+    /**
+     * Calculates the supplier's days of supply for a given material, partner, and site over a specified number of days.
+     * It combines own and reported deliveries, and production quantities to forecast the number of days the stock will last.
+     *
+     * @param material the material identifier for which the days of supply are being calculated.
+     * @param partnerBpnl The bpnl of the supplier's partner.
+     * @param siteBpns the bpns of the site where the deliveries and productions are recorded.
+     * @param numberOfDays the number of days over which the forecast should be calculated.
+     * @return a list of {@link OwnSupplierSupply} objects, each containing the calculated days of supply for a specific date.
+     */
+    public final List<OwnSupplierSupply> calculateSupplierDaysOfSupply(String material, String partnerBpnl, String siteBpns, int numberOfDays) {
         List<OwnSupplierSupply> supplierSupply = new ArrayList<>();
         LocalDate localDate = LocalDate.now();
 
@@ -126,6 +130,10 @@ public class SupplierSupplyService {
         return stream.toList();
     }
 
+    public final List<ReportedSupplierSupply> findByPartnerBpnlAndOwnMaterialNumber(String partnerBpnl, String ownMaterialNumber) {
+        return repository.findByPartner_BpnlAndMaterial_OwnMaterialNumber(partnerBpnl, ownMaterialNumber);
+    }
+
     public boolean validate(ReportedSupplierSupply daysOfSupply) {
         return 
             daysOfSupply.getPartner() != null &&
@@ -137,12 +145,16 @@ public class SupplierSupplyService {
             (daysOfSupply.getStockLocationBPNA().equals(null) || daysOfSupply.getStockLocationBPNA().equals(daysOfSupply.getStockLocationBPNS()));
     }
 
+    /**
+     * Calculates the number of days of supply based on the current stock quantity and a list of demands.
+     * @param stockQuantity Current stock amount
+     * @param demands Sublist of demands for current iteration
+     * @return The number of days of supply that the stock can cover.
+     */
     private final double getDaysOfSupply(double stockQuantity, List<Double> demands) {
         double daysOfSupply = 0;
 
-        for (int i = 0; i < demands.size(); i++) {
-            double demand = demands.get(i);
-
+        for (double demand : demands) {
             if (stockQuantity >= demand) {
                 daysOfSupply += 1;
                 stockQuantity = stockQuantity - demand;
@@ -157,6 +169,12 @@ public class SupplierSupplyService {
         return daysOfSupply;
     }
 
+    /**
+     * Merges own and reported deliveries into a single list.
+     * @param list1 Own deliveries
+     * @param list2 Reported deliveries
+     * @return a new list containing the summed delivery quantities from the input lists.
+     */
     private static List<Double> mergeDeliveries(List<Double> list1, List<Double> list2) {
         if (list1.size() != list2.size()) {
             throw new IllegalArgumentException("Lists must be of the same length");
